requirements.txt
thrift>=0.8.0
cython>=0.16
pyzmq>=2.2.0
git+http://github.com/blackwithwhite666/pyev.git#egg=pyev
git+http://github.com/blackwithwhite666/socket_zmq.git#egg=socket_zmq
#pyev>=0.8.1-4.04
setproctitle>=1.1.6
billiard>=2.7.3.10
futures>=2.1.2

